Refactoring the Datastore Page/UI codebase and fixing tests.
